+++ composeApp/src/iosMain/kotlin/com/mnessim/researchtrackerkmp/domain/data/DBFactory.ios.kt
@@ -4,6 +4,7 @@ import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.native.NativeSqliteDriver
 import com.mnessim.Database
 
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 actual class DBFactory {
     actual fun createDriver(): SqlDriver {
         return NativeSqliteDriver(Database.Schema, "terms.db")